llm-server/requirements.txt
# LLM server dependencies
flask
flask-cors
pytest
requests
gtts
python-dotenv
chromadb

# Placeholder for LLM server dependencies
# Add packages like flask, openai, etc. as needed
flask
flask-cors
